Delete initialize_heater
