+++ src/dynamics/thermal/heater.cpp
@@ -1,5 +1,6 @@
 #include "heater.hpp"
 
+#include <cassert>
 #include <cmath>
 
 using namespace std;
@@ -41,4 +42,36 @@ void Heater::AssertHeaterParams(void) {
     std::cerr << "The value is set as 0.0." << std::endl;
     power_rating_W_ = 0.0;
   }
+}
+
+/* Import heater properties by reading CSV File (heaters.csv)
+
+[File Formats of heater.csv]
+column 1: Heater_id(int, Use values larger than or equal to 1)
+column 2: Power Rating (double, [W])
+column 3: Lower threshold of control (double, [degC])
+column 4: Upper threshold of control (double, [degC])
+
+First row is for Header, data begins from the second row
+*/
+
+Heater InitHeater(const std::vector<std::string>& heater_str) {
+  using std::stod;
+  using std::stoi;
+
+  size_t heater_str_size_defined = 4;                    // Correct size of heater_str
+  assert(heater_str.size() == heater_str_size_defined);  // Check if size of heater_str is correct
+
+  size_t heater_id = 0;
+  double power_rating_W = 0;  // [W]
+
+  // Index to read from heater_str for each parameter
+  size_t index_heater_id = 0;
+  size_t index_power_rating = 1;
+
+  heater_id = stoi(heater_str[index_heater_id]);
+  power_rating_W = stod(heater_str[index_power_rating]);
+
+  Heater heater(heater_id, power_rating_W);
+  return heater;
 }
